+++ vhost/src/vhost_user/message.rs
@@ -169,6 +169,8 @@ enum_value! {
         /// After transferring state, check the backend for any errors that may have
         /// occurred during the transfer
         CHECK_DEVICE_STATE = 43,
+        /// Get shared memory regions configuration from the backend.
+        GET_SHMEM_CONFIG = 44,
     }
 }
 
@@ -687,6 +689,44 @@ impl VhostUserSingleMemoryRegion {
 // SAFETY: Safe because all fields of VhostUserSingleMemoryRegion are POD.
 unsafe impl ByteValued for VhostUserSingleMemoryRegion {}
 impl VhostUserMsgValidator for VhostUserSingleMemoryRegion {}
+
+/// Get shared memory regions configuration.
+#[repr(C)]
+#[derive(Debug, Clone, Copy)]
+pub struct VhostUserShMemConfig {
+    /// Total number of shared memory regions
+    pub nregions: u32,
+    /// Padding for correct alignment
+    padding: u32,
+    /// Size of each memory region
+    pub memory_sizes: [u64; 256],
+}
+
+impl Default for VhostUserShMemConfig {
+    fn default() -> Self {
+        Self {
+            nregions: 0,
+            padding: 0,
+            memory_sizes: [0; 256],
+        }
+    }
+}
+
+impl VhostUserShMemConfig {
+    /// Create a new instance
+    pub fn new(nregions: u32, memory: &[u64]) -> Self {
+        let memory_sizes: [u64; 256] = std::array::from_fn(|i| *memory.get(i).unwrap_or(&0));
+        Self {
+            nregions,
+            padding: 0,
+            memory_sizes,
+        }
+    }
+}
+
+// SAFETY: Safe because all fields of VhostUserShMemConfig are POD.
+unsafe impl ByteValued for VhostUserShMemConfig {}
+impl VhostUserMsgValidator for VhostUserShMemConfig {}
 
 /// Vring state descriptor.
 #[repr(C, packed)]